+++ VaccinationSystem/VaccinationSystem/Data/ApplicationDbContext.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore;
+using VaccinationSystem.Data.Classes;
 using VaccinationSystem.Data.Configurations;
 
 namespace VaccinationSystem.Data
@@ -7,10 +8,8 @@ namespace VaccinationSystem.Data
     public class ApplicationDbContext : IdentityDbContext<ApplicationUser>
     {
         public DbSet<Disease> Diseases { get; set; }
-        public DbSet<Vaccine> Vaccine { get; set; }
-        public DbSet<Vaccination> Vaccination { get; set; }
-        public DbSet<Patient> Patients { get; set; }
-        public DbSet<Doctor> Doctors { get; set; }
+        public DbSet<Vaccine> Vaccines { get; set; }
+        public DbSet<Vaccination> Vaccinations { get; set; }
 
         public ApplicationDbContext(DbContextOptions<ApplicationDbContext> options)
             : base(options)
@@ -20,12 +19,16 @@ namespace VaccinationSystem.Data
         protected override void OnModelCreating(ModelBuilder builder)
         {
             base.OnModelCreating(builder);
-            builder.ApplyConfiguration(new DiseaseSeedConfiguration());
+
             builder.ApplyConfiguration(new RoleSeedConfiguration());
             builder.ApplyConfiguration(new AdminSeedConfiguration());
             builder.ApplyConfiguration(new DoctorSeedConfiguration());
             builder.ApplyConfiguration(new PatientSeedConfiguration());
             builder.ApplyConfiguration(new UserRoleSeedConfiguration());
+
+            builder.ApplyConfiguration(new DiseaseSeedConfiguration());
+            builder.ApplyConfiguration(new VaccineSeedConfiguration());
+            builder.ApplyConfiguration(new VaccinationSeedConfiguration());
         }
     }
 }
