+++ VaccinationSystem/VaccinationSystem/Data/Classes/Administrator.cs
@@ -1,0 +1,4 @@
+﻿namespace VaccinationSystem.Data.Classes
+{
+    public class Administrator : ApplicationUser { }
+}
