+++ VaccinationSystem/VaccinationSystem/Data/Classes/ApplicationUser.cs
@@ -1,0 +1,14 @@
+﻿using Microsoft.AspNetCore.Identity;
+using Microsoft.EntityFrameworkCore;
+using System.ComponentModel.DataAnnotations;
+
+namespace VaccinationSystem.Data.Classes
+{
+    public class ApplicationUser : IdentityUser
+    {
+        [Required]
+        public string? FirstName { get; set; }
+        [Required]
+        public string? LastName { get; set; }
+    }
+}
