+++ VaccinationSystem/VaccinationSystem/Data/Classes/Patient.cs
@@ -1,9 +1,13 @@
 ﻿using Microsoft.AspNetCore.Identity;
+using System.ComponentModel.DataAnnotations;
 
-namespace VaccinationSystem.Data
+namespace VaccinationSystem.Data.Classes
 {
     public class Patient : ApplicationUser
     {
+        [Required]
+        public string? Pesel { get; set; }
+    
         // nie dodaję VaccinationList, bo to brzmi jak głupie dublowanie informacji w dwóch strukturach
 
         void RegisterVisit(Vaccination vaccination)
