+++ VaccinationSystem/VaccinationSystem/Data/Classes/Vaccination.cs
@@ -1,7 +1,7 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace VaccinationSystem.Data
+namespace VaccinationSystem.Data.Classes
 {
     public enum VaccinationStatus
     {
@@ -13,17 +13,16 @@ namespace VaccinationSystem.Data
         [Key]
         public int Id { get; set; }
 
-        [Required]
-        public string PatientId { get; set; }
+        public string? PatientId { get; set; }
 
         [ForeignKey("PatientId")]
-        public Patient Patient { get; set; }
+        public Patient? Patient { get; set; }
 
         [Required]
-        public string DoctorId { get; set; }  
+        public string? DoctorId { get; set; }  
 
         [ForeignKey("DoctorId")]
-        public Doctor Doctor { get; set; }
+        public Doctor? Doctor { get; set; }
 
         [Required]
         public VaccinationStatus Status { get; set; }
@@ -35,7 +34,7 @@ namespace VaccinationSystem.Data
         public int VaccineId { get; set; }  
 
         [ForeignKey("VaccineId")]
-        public Vaccine Vaccine { get; set; }
+        public Vaccine? Vaccine { get; set; }
 
         // AssingDoctor, AssignPatient, ChangeStatus to odpowiednie settery, więc nie ma po co tworzyć tych metod
     }
