+++ VaccinationSystem/VaccinationSystem/Data/Configurations/VaccinationSeedConfiguration.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
+using VaccinationSystem.Data.Classes;
 
 namespace VaccinationSystem.Data.Configurations
 {
@@ -29,7 +30,7 @@ namespace VaccinationSystem.Data.Configurations
                 },
                 new Vaccination
                 {
-                    Id = -1,
+                    Id = -3,
                     PatientId = "c1076fe7-abf6-420d-8810-6cb0f3a92f6a",
                     DoctorId = "f1076fe7-abf6-420d-8810-6cb0f3a92f6a",
                     Status = VaccinationStatus.Completed,
