+++ VaccinationSystem/VaccinationSystem/Pages/Shared/_LoginPartial.cshtml
@@ -1,4 +1,5 @@
 ﻿@using Microsoft.AspNetCore.Identity
+@using VaccinationSystem.Data.Classes
 @inject SignInManager<ApplicationUser> SignInManager
 @inject UserManager<ApplicationUser> UserManager
